Move `NodeId` to "aranya_fast_channels::testing::util"
